fix: horizontal scrolling janky and scrollbar ruining vertical size
This was due to how the smooth scrolling was done.

We solved this by removing smooth scrolling altogether for this one until we have a proper solution.

The scrollbar has been hidden in favour of fade/clickable buttons which communicate to users that this can be slided. It also has tooltips which would inform desktop users that they can use their mouse scroll wheel.